Crate ontio_std

Modules

abi
The abi module provides serialization and deserialization methods for different data types in the contract

console
The console module provides the debug function, which is used to print the log information in the contract and facilitate the debugging of the contract.

contract
The contract module provides the method to call the asset ont, ong on the ontology chain.

database
The database module provides the interface to save the data in the contract to the chain and query the data from the chain.

macros
Macro module provides common macro functions, such as base58! To convert base58 encoded address to address type

prelude
The prelude module provides common data types in the contract, and introduces some functions in the rust core library.

runtime
The runtime module provides an interface to interact with the chain in the contract

types
The types module provides common data types such as address, U128, hash, etc.

Functions

oom
Overrides the default oom

panic_fmt
Overrides the default panic_fmt