Struct ontio_std::abi::Sink

pub struct Sink { /* fields omitted */ }

Encoding different types of data into byte array.

Implementations

impl Sink

pub fn new(cap: usize) -> Self

Create a new sink entity, Specify initial capacity. For indefinite length parameters, the length of the parameter will be serialized first, and then the content of the parameter will be serialized.

Example

 let mut sink = Sink::new(0);
 sink.write("123");
 assert_eq!(sink.bytes(),[3,49,50,51]);

pub fn write<T: Encoder>(&mut self, val: T)

All data types that implement the encode interface can be serialized by calling the write method

Example

   let mut sink = Sink::new(0);
   let addr = Address::repeat_byte(1u8);
   sink.write(addr);
   sink.write("123");
   sink.write(123 as U128);

pub fn write_neovm_address(&mut self, address: &Address)

pub fn write_native_varuint(&mut self, val: u64)

pub fn bytes(&self) -> &[u8]

Used to get the serialized result in bytearray format

Example

#![feature(proc_macro_hygiene)]
use ontio_std::macros::base58;
use ontio_std::types::Address;
use ontio_std::abi::Sink;
const ONT_CONTRACT_ADDRESS: Address = base58!("AFmseVrdL9f9oyCzZefL9tG6UbvhUMqNMV");
let mut sink = Sink::new(0);
sink.write(&ONT_CONTRACT_ADDRESS);
assert_eq!(sink.into(), [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1].to_vec())

pub fn into(self) -> Vec<u8>

Used to get the serialized result in Vec format

Example

fn main() {

}