Module ontio_std::contract::ong

This module provides the operation API related to ong assets, such as balanceof, transfer, etc.

Functions

allowance
This method is used in conjunction with the approve method to query the number of approve

approve
from-address can allow to-address to transfer a certain amount of assets from from-address.

balance_of
Query the balance of ong assets

transfer
Transfer method of ong assets, Transfer ont assets from the from address to the to address

transfer_from
Spender transfers a certain amount of ong from from-address to to-address

transfer_multi
transfer_multi method of ong assets,Multiple transfers in one transaction