Module ontio_std::database

The database module provides the interface to save the data in the contract to the chain and query the data from the chain.

Structs

ListStore

Functions

delete

get

put