Module ontio_std::prelude

The prelude module provides common data types in the contract, and introduces some functions in the rust core library.

Re-exports

pub use crate::contract::TransferParam;

pub use crate::types::Address;

pub use crate::types::H256;

pub use crate::types::I128;

pub use crate::types::U128;

Modules

cmp
Functionality for ordering and comparison.

str
Unicode string slices.

string
A UTF-8 encoded, growable string.

vec
A contiguous growable array type with heap-allocated contents, written Vec<T>.

v1	Experimental
The first version of the prelude of alloc crate.

Macros

Clone
Derive macro generating an impl of the trait Clone.

Copy
Derive macro generating an impl of the trait Copy.

Debug
Derive macro generating an impl of the trait Debug.

Default
Derive macro generating an impl of the trait Default.

Eq
Derive macro generating an impl of the trait Eq.

Hash
Derive macro generating an impl of the trait Hash.

Ord
Derive macro generating an impl of the trait Ord.

PartialEq
Derive macro generating an impl of the trait PartialEq.

PartialOrd
Derive macro generating an impl of the trait PartialOrd.

RustcDecodable
Unstable implementation detail of the rustc compiler, do not use.

RustcEncodable
Unstable implementation detail of the rustc compiler, do not use.

assert
Asserts that a boolean expression is true at runtime.

cfg
Evaluates boolean combinations of configuration flags at compile-time.

column
Expands to the column number at which it was invoked.

compile_error
Causes compilation to fail with the given error message when encountered.

concat
Concatenates literals into a static string slice.

env
Inspects an environment variable at compile time.

file
Expands to the file name in which it was invoked.

format
Creates a String using interpolation of runtime expressions.

format_args
Constructs parameters for the other string-formatting macros.

global_allocator
Attribute macro applied to a static to register it as a global allocator.

include
Parses a file as an expression or an item according to the context.

include_bytes
Includes a file as a reference to a byte array.

include_str
Includes a utf8-encoded file as a string.

line
Expands to the line number on which it was invoked.

module_path
Expands to a string that represents the current module path.

option_env
Optionally inspects an environment variable at compile time.

stringify
Stringifies its arguments.

test
Attribute macro applied to a function to turn it into a unit test.

vec
Creates a Vec containing the arguments.

asm	Experimental
Inline assembly.

bench	Experimental
Attribute macro applied to a function to turn it into a benchmark test.

cfg_accessible	Experimental
Keeps the item it's applied to if the passed path is accessible, and removes it otherwise.

concat_idents	Experimental
Concatenates identifiers into one identifier.

format_args_nl	Experimental
Same as format_args, but adds a newline in the end.

global_asm	Experimental
Module-level inline assembly.

llvm_asm	Experimental
LLVM-style inline assembly.

log_syntax	Experimental
Prints passed tokens into the standard output.

test_case	Experimental
An implementation detail of the #[test] and #[bench] macros.

trace_macros	Experimental
Enables or disables tracing functionality used for debugging other macros.

Structs

Box
A pointer type for heap allocation.

String
A UTF-8 encoded, growable string.

Vec
A contiguous growable array type, written Vec<T> but pronounced 'vector'.

Enums

Option
The Option type. See the module level documentation for more.

Result
Result is a type that represents either success (Ok) or failure (Err).

Traits

AsMut
Used to do a cheap mutable-to-mutable reference conversion.

AsRef
Used to do a cheap reference-to-reference conversion.

Clone
A common trait for the ability to explicitly duplicate an object.

Copy
Types whose values can be duplicated simply by copying bits.

Default
A trait for giving a type a useful default value.

DoubleEndedIterator
An iterator able to yield elements from both ends.

Drop
Custom code within the destructor.

Eq
Trait for equality comparisons which are equivalence relations.

ExactSizeIterator
An iterator that knows its exact length.

Extend
Extend a collection with the contents of an iterator.

Fn
The version of the call operator that takes an immutable receiver.

FnMut
The version of the call operator that takes a mutable receiver.

FnOnce
The version of the call operator that takes a by-value receiver.

From
Used to do value-to-value conversions while consuming the input value. It is the reciprocal of Into.

Into
A value-to-value conversion that consumes the input value. The opposite of From.

IntoIterator
Conversion into an Iterator.

Iterator
An interface for dealing with iterators.

Ord
Trait for types that form a total order.

PartialEq
Trait for equality comparisons which are partial equivalence relations.

PartialOrd
Trait for values that can be compared for a sort-order.

Send
Types that can be transferred across thread boundaries.

Sized
Types with a constant size known at compile time.

Sync
Types for which it is safe to share references between threads.

ToString
A trait for converting a value to a String.

Unpin
Types that can be safely moved after being pinned.

Functions

drop
Disposes of a value.