Module ontio_std::runtime

The runtime module provides an interface to interact with the chain in the contract

Functions

address
Get the address of current executing contract

block_height
Get current block height

call_contract
With this method, another contract can be called, When calling Neo contract and native contract across contracts, the parameter construction is slightly different. Please refer to the corresponding examples.

caller
return Caller's contract address

check_witness
Check signature

contract_create
Create a new contract based on the parameters passed in

contract_delete
delete the contract

contract_migrate
contract migrate

current_blockhash
return current block hash

current_txhash
return current tx hash

entry_address
return the entry address

input
Get input data from transaction or caller contract

panic
When the function is executed, all writes to the chain will be cancelled, and the error message will be returned.

ret
return the result of execution and exit contract execution

sha256
Calculate the hash value

storage_delete
Delete key-value pairs according to the key

storage_read
Read key-value pairs according to key.

storage_write
Save key-value as a key-value pair

timestamp
Get timestamp in current block